Displays a correlation matrix of the quantitative data of a numeric matrix.

xxxx

Usage

mod_ds_metacell_Histos_ui(id)

mod_ds_metacell_Histos_server(
  id,
  dataIn = reactive({
     NULL
 }),
  group = reactive({
     NULL
 }),
  pal = reactive({
     NULL
 }),
  pattern = reactive({
     NULL
 }),
  showSelect = reactive({
     TRUE
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

mod_ds_metacell_Histos(obj, group, pal, pattern = NULL, showSelect)

Arguments

id

xxx

dataIn

xxx

group

xxx

pal

xxx

pattern

xxx

showSelect

xxx

remoteReset

A logical(1) which acts as a remote command to reset the module to its default values. Default is FALSE.

is.enabled

xxx

obj

An instance of the class SummarizedExperiment

Value

NA

Examples

if (interactive()){
data(Exp1_R25_prot, package = "DaparToolshedData")
grp <- DaparToolshed::design_qf(Exp1_R25_prot)$Condition
shiny::runApp(mod_ds_metacell_Histos(Exp1_R25_prot[[1]], group = grp))

# Test with pattern already defined
pat <- c("Missing MEC", "Missing POV")
shiny::runApp(mod_ds_metacell_Histos(Exp1_R25_prot[[1]], pattern = pat, group = grp))
}