Volcanoplot of the differential analysis

#' Plots an interactive volcanoplot after the differential analysis. Typically, the log of Fold Change is represented on the X-axis and the log10 of the p-value is drawn on the Y-axis. When the th_pval and the th_logfc are set, two lines are drawn respectively on the y-axis and the X-axis to visually distinguish between differential and non differential data. With the use of the package Highcharter, a customizable tooltip appears when the user put the mouse's pointer over a point of the scatter plot.

Usage

diffAnaVolcanoplot_rCharts(
  df,
  th_pval = 1e-60,
  th_logfc = 0,
  conditions = NULL,
  clickFunction = NULL,
  pal = NULL
)

Arguments

df

A dataframe which contains the following slots : x : a vector of the log(fold change) values of the differential analysis, y : a vector of the p-value values returned by the differential analysis. index : a vector of the rowanmes of the data. This dataframe must has been built with the option stringsAsFactors set to FALSE. There may be additional slots which will be used to show informations in the tooltip. The name of these slots must begin with the prefix "tooltip_". It will be automatically removed in the plot.

th_pval

A floating number which represents the p-value that separates differential and non-differential data.

th_logfc

A floating number which represents the log of the Fold Change that separates differential and non-differential data.

conditions

A list of the names of condition 1 and 2 used for the differential analysis.

clickFunction

A string that contains a JavaScript function used to show info from slots in df. The variable this.index refers to the slot named index and allows to retrieve the right row to show in the tooltip.

pal

xxx

Value

An interactive volcanoplot

Author

Samuel Wieczorek

Examples

if (interactive()){
library(highcharter)
library(DaparToolshed)
library(SummarizedExperiment)
data(Exp1_R25_prot, package = "DaparToolshedData")
obj <- Exp1_R25_prot
# Simulate imputation of missing values
obj <- DaparToolshed::NAIsZero(obj, 1)
obj <- DaparToolshed::NAIsZero(obj, 2)
qData <- as.matrix(SummarizedExperiment::assay(obj[[2]]))
sTab <- colData(obj[[2]])
limma <- limmaCompleteTest(qData, sTab)

df <- data.frame(
  x = limma$logFC[["25fmol_vs_10fmol_logFC"]],
  y = -log10(limma$P_Value[["25fmol_vs_10fmol_pval"]]),
  index = as.character(rownames(obj[[2]]))
)
colnames(df) <- c("x", "y", "index")
tooltipSlot <- c("Fasta_headers", "Sequence_length")
df <- cbind(df, colData(obj[[2]])[, tooltipSlot])
colnames(df) <- gsub(".", "_", colnames(df), fixed = TRUE)
if (ncol(df) > 3) {
  colnames(df)[seq.int(from = 4, to = ncol(df))] <-
    paste("tooltip_", colnames(df)[seq.int(from = 4, to = ncol(df))],
      sep = ""
    )
}
hc_clickFunction <- JS("function(event) {
Shiny.onInputChange('eventPointClicked',
[this.index]+'_'+ [this.series.name]);}")
cond <- c("25fmol", "10fmol")
diffAnaVolcanoplot_rCharts(
  df,
  th_pval = 2.5,
  th_logfc = 1,
  conditions = cond,
  clickFunction = hc_clickFunction
)
}